chore(core-manager): inherit GenerateLogGz from GenerateLog
