+++ packages/core-manager/src/workers/actions/generate-log.ts
@@ -1,12 +1,7 @@
 import dayjs from "dayjs";
 import utc from "dayjs/plugin/utc";
-import { createWriteStream, ensureDirSync, renameSync } from "fs-extra";
-import { dirname, join } from "path";
-import { Writable } from "stream";
-import zlib from "zlib";
 
 import { Database, Schema } from "../../database/database";
-import { LogsResult } from "../../database/logs-database-service";
 
 dayjs.extend(utc);
 
@@ -19,54 +14,13 @@ export interface Options {
 }
 
 export class GenerateLog {
-    private readonly database: Database;
+    protected readonly database: Database;
 
-    public constructor(private readonly options: Options) {
+    public constructor(protected readonly options: Options) {
         this.database = new Database(options.databaseFilePath, options.schema);
 
         this.database.boot();
     }
 
-    public async execute(): Promise<void> {
-        const iterator = this.database.getAllIterator("logs", this.options.query);
-
-        const stream = this.prepareOutputStream();
-
-        for (const log of iterator) {
-            stream.write(this.formatLog(log));
-        }
-
-        await new Promise((resolve) => {
-            stream.end(() => {
-                resolve();
-            });
-        });
-
-        ensureDirSync(dirname(this.getFilePath()));
-        renameSync(this.getTempFilePath(), this.getFilePath());
-    }
-
-    private formatLog(log: LogsResult): string {
-        return `[${dayjs.unix(log.timestamp).utc().format("YYYY-MM-DD HH:mm:ss.SSS")}] ${log.level.toUpperCase()} : ${
-            log.content
-        }\n`;
-    }
-
-    private getFilePath(): string {
-        return join(process.env.CORE_PATH_DATA!, "log-archive", this.options.logFileName);
-    }
-
-    private getTempFilePath(): string {
-        return join(process.env.CORE_PATH_TEMP!, "log-archive", this.options.logFileName);
-    }
-
-    private prepareOutputStream(): Writable {
-        ensureDirSync(dirname(this.getTempFilePath()));
-
-        const stream = zlib.createGzip();
-
-        stream.pipe(createWriteStream(this.getTempFilePath()));
-
-        return stream;
-    }
+    public async execute(): Promise<void> {}
 }
