+++ packages/core-manager/src/workers/worker.ts
@@ -1,9 +1,10 @@
 import { workerData } from "worker_threads";
 
-import { GenerateLog, Options as GenerateLogOptions } from "./actions/generate-log";
+import { Options as GenerateLogOptions } from "./actions/generate-log";
+import { GenerateLogGz } from "./actions/generate-log-gz";
 
 const main = async () => {
-    const action = new GenerateLog(workerData as GenerateLogOptions);
+    const action = new GenerateLogGz(workerData as GenerateLogOptions);
 
     await action.execute();
 };
